More flexible helper functions in testGitsummary.py

+++ testGitsummary.py
@@ -56,16 +56,24 @@ def rmtreeErrorHandler(func, path, exception):
 #-----------------------------------------------------------------------------
 # Helpers
 #-----------------------------------------------------------------------------
-def createAndCommitFile(filename, commitMsg = 'Commit Message'):
+def createAndCommitFile(
+    filename,
+    contents = 'Default contents',
+    commitMsg = 'Commit message'
+):
     """
-    Create the specified file (empty) in the current working directory then
-    'git add' and 'git commit'.
+    Create the specified file with the specified contents in the current working
+    directory then 'git add' and 'git commit'.
+
+    An error will be thrown if the file exists already.
 
     Args
         String filename  - The name of the file to create
+        String contents  - The contents to be written to the file
         String commitMsg - The commit message to use
     """
-    newFile = open(filename, 'w')
+    newFile = open(filename, 'x')
+    newFile.write(contents)
     newFile.close()
     execute(['git', 'add', filename])
     execute(['git', 'commit', '-m', commitMsg])
@@ -134,15 +142,19 @@ def modifyAndCommitFile(
     commitMsg = 'Default commit message'
 ):
     """
-    Modify the specified file (creating it if it doesn't exist), with the
-    specified contents, in the current working directory then 'git add' and
-    'git commit'.
+    Replace the contents of the specified file with the specified contents, in
+    the current working directory then 'git add' and 'git commit'.
+
+    Throws an error if the file does not already exist.
 
     Args
         String filename  - The name of the file
-        String contents  - The contents for the file
+        String contents  - The contents to be written to the file
         String commitMsg - The commit message to use
     """
+    if (not os.path.isfile(filename)):
+        raise Exception('File does not exist')
+
     modifiedFile = open(filename, 'w')
     modifiedFile.write(contents)
     modifiedFile.close()
@@ -353,7 +365,7 @@ class Test_gitGetCommitDetails(unittest.TestCase):
         COMMIT_MSG = 'This is the message'
 
         createNonEmptyGitRepository()
-        createAndCommitFile('newFile', COMMIT_MSG)
+        createAndCommitFile('newFile', '', COMMIT_MSG)
 
         # rev-list output will be:
         # commit [fullHash]
@@ -784,12 +796,12 @@ class Test_gitGetFileStatuses(unittest.TestCase):
         # Make the changes in BRANCH1
         execute(['git', 'checkout', '-b', BRANCH1, 'master'])
         modifyAndCommitFile(testFile1, 'abcde')
-        modifyAndCommitFile(testFile2, 'abcde')
+        createAndCommitFile(testFile2, 'abcde')
 
         # Make the changes in BRANCH2
         execute(['git', 'checkout', '-b', BRANCH2, 'master'])
         modifyAndCommitFile(testFile1, 'fghij')
-        modifyAndCommitFile(testFile2, 'fghij')
+        createAndCommitFile(testFile2, 'fghij')
 
         # Merge BRANCH1 into BRANCH2, thereby causing the merge conflicts.
         # Can't use execute() helper since 'git pull' will return a non-zero
